refactored component structure to be in line with team outline
