+++ backend/db/config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
-    "username": "",
-    "password": "",
+    "username": "labber",
+    "password": "labber",
     "database": "final_project",
     "host": "127.0.0.1",
     "dialect": "postgres",
