+++ backend/server.js
@@ -37,7 +37,7 @@ app.use(function(err, req, res, next) {
   // set locals, only providing error in development
   res.locals.message = err.message;
   res.locals.error = req.app.get('env') === 'development' ? err : {};
-
+console.log(err.status)
   // render the error page
   res.status(err.status || 500);
