+++ client/src/components/QuestionAnswer/styles.scss
@@ -44,7 +44,7 @@ $prompt-height: 1rem;
     &--create {
       background-color: rgba($booked-color, $active-opacity);
       display: block;
-      border: 5px solid $indezone-navy;
+      border: 3px solid $indezone-yellow;
     }
 
     &--show {
